requirements.txt
Pillow
requests
ics
pytz
PyYAML